+++ docs/_config.js
@@ -10,4 +10,7 @@ site.copy("logo.svg");
 site.copy("favicon.ico");
 site.copy("favicon-32x32.png");
 
+site.script("listar", "ls -al _site", "ls -al .");
+site.addEventListener("beforeBuild", "listar");
+
 export default site;