+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.8.2] - Unreleased
+## [0.9.0] - 2020-11-04
 ### Added
 - New method `script()` to execute scripts like a task manager
 - Allow to run scripts in events
@@ -159,7 +159,7 @@ First version
 
 [#1]: https://github.com/oscarotero/lume/issues/1
 
-[0.8.2]: https://github.com/oscarotero/lume/compare/v0.8.1...HEAD
+[0.9.0]: https://github.com/oscarotero/lume/compare/v0.8.1...v0.9.0
 [0.8.1]: https://github.com/oscarotero/lume/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/oscarotero/lume/compare/v0.7.3...v0.8.0
 [0.7.3]: https://github.com/oscarotero/lume/compare/v0.7.2...v0.7.3